# Better altitude fetches in shader