Closes #8

The new `get_surface_alt` is highly optimized to only care about altitudes. It attempts to minimize the dependencies between instructions. Currently only affects regular ray tracing.
Rough measurement show about 10% performance boost in ray tracing in default settings (5.5ms to 5.0ms per frame), and 25% in double-quality settings (8.3ms to 6.5ms).