@kvark 2022-02-04
Second commits actually implements the smooth sampling (possibly, wrong?) but doesn't enable it.
Here is the original render:
![filtered-terrain-base](https://user-images.githubusercontent.com/107301/152568918-b173e2cf-f73d-4d0f-821a-b447057b7d8e.jpg)
Here is the smoothed one:
![filtered-terrain](https://user-images.githubusercontent.com/107301/152568951-42c9343a-bbed-46d6-a8fe-6ab421c7ac72.jpg)