@kvark 2022-02-04
Ok, that one wasn't entirely correct. Here is the fixed one:
![filtered-terrain](https://user-images.githubusercontent.com/107301/152588189-75993b2e-22f1-4b1c-9e0d-a56376ba422e.jpg)